chore: rename config.ts to setup.ts
