+++ src/typing/global.d.ts
@@ -1,4 +1,4 @@
-import { DefaultNetwork } from "config/config";
+import { DefaultNetwork } from "config/setup";
 declare global {
   interface Window {
     ethereum: any;
