+++ src/ui/modals/AddApproveArbitrator.tsx
@@ -17,7 +17,7 @@ import {
   startListeningNetwork,
   switchNetwork,
 } from "wallet";
-import { DefaultNetwork } from "config/config";
+import { DefaultNetwork } from "config/setup";
 import { IncorrectNetwork } from "ui/components/IncorrectNetwork";
 
 /**
