+++ src/ui/modals/Challenge.tsx
@@ -34,7 +34,7 @@ import {
   startListeningNetwork,
   switchNetwork,
 } from "wallet";
-import { DefaultNetwork } from "config/config";
+import { DefaultNetwork } from "config/setup";
 import { IncorrectNetwork } from "ui/components/IncorrectNetwork";
 import { SpinnerIcon } from "../components/icons/Spinner";
