+++ src/ui/modals/Claim.tsx
@@ -22,7 +22,7 @@ import {
   startListeningNetwork,
   switchNetwork,
 } from "wallet";
-import { DefaultNetwork } from "config/config";
+import { DefaultNetwork } from "config/setup";
 import { IncorrectNetwork } from "ui/components/IncorrectNetwork";
 
 type IBalanceWithTokenUSD = IBalanceWithTokenInfo & {
