+++ src/ui/modals/Release.tsx
@@ -26,7 +26,7 @@ import {
   startListeningNetwork,
   switchNetwork,
 } from "wallet";
-import { DefaultNetwork } from "config/config";
+import { DefaultNetwork } from "config/setup";
 import { IncorrectNetwork } from "ui/components/IncorrectNetwork";
 
 type IProtectedActions = {
