+++ src/ui/modals/SingleClaim.tsx
@@ -19,7 +19,7 @@ import {
   getExchangeRates,
 } from "../../helpers";
 import { isCorrectNetworkConnected, switchNetwork } from "wallet";
-import { DefaultNetwork } from "config/config";
+import { DefaultNetwork } from "config/setup";
 import { IncorrectNetwork } from "ui/components/IncorrectNetwork";
 
 type IBalanceWithTokenUSD = IBalanceWithTokenInfo & {
